+++ deepomatic/cli/cmds/platform/__init__.py
@@ -9,4 +9,3 @@ class PlatformCommand(Command):
     from .app import AppCommand
     from .app_version import AppVersionCommand
     from .service import ServiceCommand
-    from .site import SiteCommand
