+++ tests/test_site.py
@@ -1,7 +1,10 @@
 import os
 from uuid import uuid4
 from deepomatic.cli.lib.site import SiteManager
+from deepomatic.api.client import Client
 from contextlib import contextmanager
+from test_platform import (app_version, call_deepo,
+                           deploy_api_key, deploy_api_url)
 import tempfile
 import shutil
 
@@ -102,7 +105,7 @@ class MockApi(object):
 
     def get(self, path):
         site_id, extra_path = self.get_site_id_from_path(path)
-        if extra_path and extra_path[0] == 'docker-compose':
+        if extra_path and 'docker-compose' in extra_path[0]:
             return b'docker-compose'
         return self.sites.get(site_id)
 
@@ -139,6 +142,17 @@ def setup():
         yield manager
     finally:
         shutil.rmtree(tmp_dir, ignore_errors=True)
+
+
+@contextmanager
+def site():
+    with app_version() as (app_version_id, app_id):
+        args = "site create -n test_si -d xyz -v {}".format(app_version_id)
+        result = call_deepo(args)
+        _, site_id = result.split(':')
+
+        yield site_id.strip(), app_version_id, app_id
+        call_deepo("site delete --id {}".format(site_id))
 
 
 class TestSite(object):
@@ -213,3 +227,45 @@ class TestSite(object):
             assert(site_id in manager.list())
             manager.uninstall(site_id)
             assert(site_id not in manager.list())
+
+    def test_site(self):
+        with app_version() as (app_version_id, app_id):
+            args = "site create -n test_si -d xyz -v {}".format(app_version_id)
+            result = call_deepo(args)
+            message, site_id = result.split(':')
+            assert message == 'New site created with id'
+
+            args = "site update --id {} --app_version_id {}".format(site_id, app_version_id)
+            message = call_deepo(args)
+            assert message == 'Site{} updated'.format(site_id)
+
+            args = "site delete --id {}".format(site_id)
+            message = call_deepo(args)
+            assert message == 'Site{} deleted'.format(site_id)
+
+    def test_site_deployment_manifest(self):
+        with site() as (site_id, app_version_id, app_id):
+            # create services
+            for service in ['worker-nn', 'workflow-server', 'customer-api']:
+                call_deepo("platform service create -a {} -n {}".format(app_id, service))
+
+            client = Client(api_key=deploy_api_key, host=deploy_api_url)
+            client.http_helper.post('/accounts/me/read-only-keys',
+                                    data={'name': 'SITE_0-{}-test-deepocli'.format(site_id)})
+            args = "site manifest -i {} -t docker-compose".format(site_id)
+            message = call_deepo(args)
+            assert message.startswith('version: "2.4"')
+            assert 'services:' in message
+            assert 'neural-worker:' in message
+            assert 'workflow-server:' in message
+            assert 'customer-api:' in message
+
+            args = "site manifest -i {} -t gke".format(site_id)
+            message = call_deepo(args)
+            assert message.startswith('apiVersion: apps/v1')
+            assert 'kind: StatefulSet' in message
+            assert 'containers:' in message
+            assert '- name: neural-worker' in message
+            assert '- name: workflow-server' in message
+            assert '- name: customer-api' in message
+            assert 'kind: Ingress' in message
